requirements.txt
aiohttp
uvloop
cachetools
hikari
orjson
pyyaml
